Refactoring Message Parser main.py
