+++ containers/message-refiner/assets/section_loincs.json
@@ -1,0 +1,83 @@
+{
+  "46240-8": {
+    "details": [
+      "History of encounters",
+      "2.16.840.1.113883.10.20.22.2.22.1",
+      "2015-08-01",
+      "Encounters"
+    ],
+    "required": true
+  },
+  "10164-2": {
+    "details": [
+      "History of Present Illness",
+      "2.16.840.1.113883.10.20.22.2.20",
+      "2015-08-01",
+      "History of Present Illness"
+    ],
+    "required": true
+  },
+  "29549-3": {
+    "details": [
+      "Medications Administered",
+      "2.16.840.1.113883.10.20.22.2.38",
+      "2014-06-09",
+      "Medications Administered"
+    ],
+    "required": true
+  },
+  "18776-5": {
+    "details": [
+      "Plan of Treatment",
+      "2.16.840.1.113883.10.20.22.2.10",
+      "2014-06-09",
+      "Plan of Treatment"
+    ],
+    "required": true
+  },
+  "11450-4": {
+    "details": [
+      "Problem List",
+      "2.16.840.1.113883.10.20.22.2.5.1",
+      "2015-08-01",
+      "Problem List"
+    ],
+    "required": true
+  },
+  "29299-5": {
+    "details": [
+      "Reason For Visit",
+      "2.16.840.1.113883.10.20.22.2.12",
+      "2015-08-01",
+      "Reason For Visit"
+    ],
+    "required": true
+  },
+  "30954-2": {
+    "details": [
+      "Relevant diagnostic tests and/or laboratory data",
+      "2.16.840.1.113883.10.20.22.2.3.1",
+      "2015-08-01",
+      "Results"
+    ],
+    "required": true
+  },
+  "29762-2": {
+    "details": [
+      "Social History",
+      "2.16.840.1.113883.10.20.22.2.17",
+      "2015-08-01",
+      "Social History"
+    ],
+    "required": true
+  },
+  "11369-6": {
+    "details": [
+      "History of Immunizations",
+      "2.16.840.1.113883.10.20.22.4.52",
+      "2015-08-01",
+      "Immunizations"
+    ],
+    "required": false
+  }
+}
